+++ cli_example_test.go
@@ -3,7 +3,7 @@ package goyek_test
 import "github.com/goyek/goyek"
 
 func Example() {
-	flow := goyek.New()
+	flow := &goyek.Taskflow{}
 
 	task1 := flow.Register(goyek.Task{
 		Name:    "task-1",