+++ parameter_test.go
@@ -39,9 +39,10 @@ func Test_bool_param(t *testing.T) {
 	for index, tc := range tt {
 		tc := tc
 		t.Run("case "+strconv.Itoa(index), func(t *testing.T) {
-			flow := goyek.New()
-			param := flow.RegisterBoolParam(tc.defaultValue, goyek.ParamInfo{
-				Name: "b",
+			flow := &goyek.Taskflow{}
+			param := flow.RegisterBoolParam(goyek.BoolParam{
+				Name:    "b",
+				Default: tc.defaultValue,
 			})
 			var got bool
 			exitCode := runTaskflowWith(flow, param, func(tf *goyek.TF) { got = param.Get(tf) }, tc.args)
@@ -53,9 +54,10 @@ func Test_bool_param(t *testing.T) {
 }
 
 func Test_bool_param_help(t *testing.T) {
-	flow := goyek.New()
-	param := flow.RegisterBoolParam(true, goyek.ParamInfo{
-		Name: "bool",
+	flow := &goyek.Taskflow{}
+	param := flow.RegisterBoolParam(goyek.BoolParam{
+		Name:    "bool",
+		Default: true,
 	})
 	exitCode := runTaskflowWith(flow, param, func(tf *goyek.TF) {}, []string{"-h"})
 
@@ -82,9 +84,10 @@ func Test_int_param(t *testing.T) {
 	for index, tc := range tt {
 		tc := tc
 		t.Run("case "+strconv.Itoa(index), func(t *testing.T) {
-			flow := goyek.New()
-			param := flow.RegisterIntParam(tc.defaultValue, goyek.ParamInfo{
-				Name: "i",
+			flow := &goyek.Taskflow{}
+			param := flow.RegisterIntParam(goyek.IntParam{
+				Name:    "i",
+				Default: tc.defaultValue,
 			})
 			var got int
 			exitCode := runTaskflowWith(flow, param, func(tf *goyek.TF) { got = param.Get(tf) }, tc.args)
@@ -96,9 +99,10 @@ func Test_int_param(t *testing.T) {
 }
 
 func Test_int_param_help(t *testing.T) {
-	flow := goyek.New()
-	param := flow.RegisterIntParam(123, goyek.ParamInfo{
-		Name: "int",
+	flow := &goyek.Taskflow{}
+	param := flow.RegisterIntParam(goyek.IntParam{
+		Name:    "int",
+		Default: 123,
 	})
 	exitCode := runTaskflowWith(flow, param, func(tf *goyek.TF) {}, []string{"-h"})
 
@@ -122,9 +126,10 @@ func Test_string_param(t *testing.T) {
 	for index, tc := range tt {
 		tc := tc
 		t.Run("case "+strconv.Itoa(index), func(t *testing.T) {
-			flow := goyek.New()
-			param := flow.RegisterStringParam(tc.defaultValue, goyek.ParamInfo{
-				Name: "s",
+			flow := &goyek.Taskflow{}
+			param := flow.RegisterStringParam(goyek.StringParam{
+				Name:    "s",
+				Default: tc.defaultValue,
 			})
 			var got string
 			exitCode := runTaskflowWith(flow, param, func(tf *goyek.TF) { got = param.Get(tf) }, tc.args)
@@ -136,9 +141,10 @@ func Test_string_param(t *testing.T) {
 }
 
 func Test_string_param_help(t *testing.T) {
-	flow := goyek.New()
-	param := flow.RegisterStringParam("abc", goyek.ParamInfo{
-		Name: "string",
+	flow := &goyek.Taskflow{}
+	param := flow.RegisterStringParam(goyek.StringParam{
+		Name:    "string",
+		Default: "abc",
 	})
 	exitCode := runTaskflowWith(flow, param, func(tf *goyek.TF) {}, []string{"-h"})