+++ taskflow_test.go
@@ -35,7 +35,7 @@ func Test_Register_errors(t *testing.T) {
 	}
 	for _, tc := range testCases {
 		t.Run(tc.desc, func(t *testing.T) {
-			flow := goyek.New()
+			flow := &goyek.Taskflow{}
 
 			act := func() { flow.Register(tc.task) }
 
@@ -229,7 +229,7 @@ func Test_empty_command(t *testing.T) {
 }
 
 func Test_invalid_args(t *testing.T) {
-	flow := goyek.New()
+	flow := &goyek.Taskflow{}
 	flow.Register(goyek.Task{
 		Name: "task",
 	})
@@ -260,8 +260,8 @@ func Test_invalid_args(t *testing.T) {
 }
 
 func Test_help(t *testing.T) {
-	flow := goyek.New()
-	fastParam := flow.RegisterBoolParam(false, goyek.ParamInfo{
+	flow := &goyek.Taskflow{}
+	fastParam := flow.RegisterBoolParam(goyek.BoolParam{
 		Name:  "fast",
 		Usage: "simulates fast-lane processing",
 	})
@@ -345,7 +345,7 @@ func Test_concurrent_printing(t *testing.T) {
 			args = append(args, "task")
 			exitCode := flow.Run(context.Background(), args...)
 
-			assertEqual(t, exitCode, goyek.CodeFailure, "should fail")
+			assertEqual(t, exitCode, goyek.CodeFail, "should fail")
 			assertContains(t, sb.String(), "from child goroutine", "should contain log from child goroutine")
 			assertContains(t, sb.String(), "from main goroutine", "should contain log from main goroutine")
 		})
@@ -385,18 +385,22 @@ func (value *arrayValue) String() string {
 func (value *arrayValue) IsBool() bool { return false }
 
 func Test_params(t *testing.T) {
-	flow := goyek.New()
-	boolParam := flow.RegisterBoolParam(true, goyek.ParamInfo{
-		Name: "b",
+	flow := &goyek.Taskflow{}
+	boolParam := flow.RegisterBoolParam(goyek.BoolParam{
+		Name:    "b",
+		Default: true,
 	})
-	intParam := flow.RegisterIntParam(1, goyek.ParamInfo{
-		Name: "i",
+	intParam := flow.RegisterIntParam(goyek.IntParam{
+		Name:    "i",
+		Default: 1,
 	})
-	stringParam := flow.RegisterStringParam("abc", goyek.ParamInfo{
-		Name: "s",
+	stringParam := flow.RegisterStringParam(goyek.StringParam{
+		Name:    "s",
+		Default: "abc",
 	})
-	arrayParam := flow.RegisterValueParam(func() goyek.ParamValue { return &arrayValue{} }, goyek.ParamInfo{
-		Name: "array",
+	arrayParam := flow.RegisterValueParam(goyek.ValueParam{
+		Name:     "array",
+		NewValue: func() goyek.ParamValue { return &arrayValue{} },
 	})
 	var gotBool bool
 	var gotInt int
@@ -428,7 +432,7 @@ func Test_params(t *testing.T) {
 }
 
 func Test_invalid_params(t *testing.T) {
-	flow := goyek.New()
+	flow := &goyek.Taskflow{}
 	flow.Register(goyek.Task{
 		Name:    "task",
 		Command: func(tf *goyek.TF) {},
@@ -440,28 +444,33 @@ func Test_invalid_params(t *testing.T) {
 }
 
 func Test_unused_params(t *testing.T) {
-	flow := goyek.New()
+	flow := &goyek.Taskflow{}
 	flow.DefaultTask = flow.Register(goyek.Task{Name: "task", Command: func(tf *goyek.TF) {}})
-	flow.RegisterBoolParam(false, goyek.ParamInfo{Name: "unused"})
+	flow.RegisterBoolParam(goyek.BoolParam{Name: "unused"})
 
 	assertPanics(t, func() { flow.Run(context.Background()) }, "should fail because of unused parameter")
 }
 
 func Test_param_registration_error_empty_name(t *testing.T) {
-	flow := goyek.New()
-	assertPanics(t, func() { flow.RegisterBoolParam(false, goyek.ParamInfo{Name: ""}) }, "empty name")
+	flow := &goyek.Taskflow{}
+	assertPanics(t, func() { flow.RegisterBoolParam(goyek.BoolParam{Name: ""}) }, "empty name")
+}
+
+func Test_param_registration_error_no_default(t *testing.T) {
+	flow := &goyek.Taskflow{}
+	assertPanics(t, func() { flow.RegisterValueParam(goyek.ValueParam{Name: "custom"}) }, "custom parameter must have default value factory")
 }
 
 func Test_param_registration_error_double_name(t *testing.T) {
-	flow := goyek.New()
-	info := goyek.ParamInfo{Name: "double"}
-	flow.RegisterBoolParam(false, info)
-	assertPanics(t, func() { flow.RegisterBoolParam(false, info) }, "double name")
+	flow := &goyek.Taskflow{}
+	name := "double"
+	flow.RegisterStringParam(goyek.StringParam{Name: name})
+	assertPanics(t, func() { flow.RegisterBoolParam(goyek.BoolParam{Name: name}) }, "double name")
 }
 
 func Test_unregistered_params(t *testing.T) {
-	foreignParam := goyek.New().RegisterBoolParam(false, goyek.ParamInfo{Name: "foreign"})
-	flow := goyek.New()
+	foreignParam := (&goyek.Taskflow{}).RegisterBoolParam(goyek.BoolParam{Name: "foreign"})
+	flow := &goyek.Taskflow{}
 	flow.Register(goyek.Task{
 		Name: "task",
 		Command: func(tf *goyek.TF) {
@@ -471,11 +480,11 @@ func Test_unregistered_params(t *testing.T) {
 
 	exitCode := flow.Run(context.Background(), "task")
 
-	assertEqual(t, goyek.CodeFailure, exitCode, "should fail because of unregistered parameter")
+	assertEqual(t, goyek.CodeFail, exitCode, "should fail because of unregistered parameter")
 }
 
 func Test_defaultTask(t *testing.T) {
-	flow := goyek.New()
+	flow := &goyek.Taskflow{}
 	taskRan := false
 	task := flow.Register(goyek.Task{
 		Name: "task",